fix: Update version of Nvidia driver in cloud images.
Upgrade from 410.104 to 410.129; 410.129 is apparently the most recent
driver version that still works with CUDA 10.0.

Along the way, switch from the "XFree86" package to the "Tesla" package.
It is a little unclear what the difference is, but the Nvidia download
site guides you to the latter when you're using enterprise GPUs, which
should always be the case in public cloud environments.